requirements.txt
streamlit
pandas
plotly
openpyxl
google-auth
google-api-python-client

